Update flat parser to only return path
