+++ tests/Router/Parsers/FlatTest.php
@@ -1,0 +1,26 @@
+<?php
+/**
+ * @author Dimitar Dimitrov <daghostman.dd@gmail.com>
+ */
+
+namespace Tests\Router\Parsers;
+
+use Onion\Framework\Interfaces\Router\ParserInterface;
+use Onion\Framework\Router\Parsers\Flat;
+
+class FlatTest extends \PHPUnit_Framework_TestCase
+{
+    /**
+     * @var ParserInterface
+     */
+    protected $parser;
+    public function setUp()
+    {
+        $this->parser = new Flat();
+    }
+
+    public function testPathParsing()
+    {
+        $this->assertSame('/', $this->parser->parse('/'));
+    }
+}
